apiEndi/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.1
mysqlclient==2.0.1
netifaces==0.10.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
socketio==0.2.1
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.2.0
